+++ monosize.config.mjs
@@ -1,4 +1,6 @@
 // @ts-check
+import fs from 'node:fs';
+import path from 'node:path';
 import webpackBundler from 'monosize-bundler-webpack';
 import createAzureStorage from 'monosize-storage-azure';
 
@@ -11,6 +13,13 @@ const config = {
   bundler: webpackBundler(config => {
     return config;
   }),
+  reportResolvers: {
+    packageName: async packageRoot => {
+      const packageJson = await fs.promises.readFile(path.join(packageRoot, 'package.json'), 'utf-8');
+      const json = JSON.parse(packageJson);
+      return json.name.replace('@fluentui/', '');
+    },
+  },
 };
 
 export default config;