+++ packages/react-components/babel-preset-global-context/src/index.ts
@@ -1,6 +1,7 @@
 import { transformPlugin } from './transformPlugin';
+import type { PluginItem } from '@babel/core';
 
-export default function preset() {
+export default function preset(): PluginItem {
   return {
     plugins: [[transformPlugin]],
   };