+++ packages/react-components/react-calendar-compat/src/components/CalendarDay/useCalendarDayStyles.styles.ts
@@ -70,12 +70,16 @@ const useMonthAndYearStyles = makeStyles({
   },
   headerIsClickable: {
     '&:hover': {
-      backgroundColor: tokens.colorBrandBackground2,
-      color: tokens.colorNeutralForeground1Hover,
+      backgroundColor: tokens.colorBrandBackgroundInvertedHover,
+      color: tokens.colorBrandForegroundOnLightHover,
       cursor: 'pointer',
+      ...shorthands.outline('1px', 'solid', tokens.colorTransparentStroke),
     },
     '&:hover:active': {
-      backgroundColor: tokens.colorBrandBackground2,
+      backgroundColor: tokens.colorBrandBackgroundInvertedPressed,
+      color: tokens.colorBrandForegroundOnLightPressed,
+      cursor: 'pointer',
+      ...shorthands.outline('1px', 'solid', tokens.colorTransparentStroke),
     },
   },
 });