+++ packages/react-components/react-nav-preview/src/components/NavItem/useNavItem.ts
@@ -1,7 +1,9 @@
 import * as React from 'react';
-import { getIntrinsicElementProps, slot, useEventCallback, mergeCallbacks } from '@fluentui/react-utilities';
+import { getIntrinsicElementProps, slot, useEventCallback, isHTMLElement } from '@fluentui/react-utilities';
+import { useARIAButtonProps } from '@fluentui/react-aria';
 import { useNavContext_unstable } from '../NavContext';
 
+import type { ARIAButtonSlotProps } from '@fluentui/react-aria';
 import type { NavItemProps, NavItemState } from './NavItem.types';
 
 /**
@@ -13,17 +15,40 @@ import type { NavItemProps, NavItemState } from './NavItem.types';
  * @param props - props from this instance of NavItem
  * @param ref - reference to root HTMLAnchorElement of NavItem
  */
-export const useNavItem_unstable = (props: NavItemProps, ref: React.Ref<HTMLAnchorElement>): NavItemState => {
-  const { onClick, value, icon } = props;
+export const useNavItem_unstable = (
+  props: NavItemProps,
+  ref: React.Ref<HTMLButtonElement | HTMLAnchorElement>,
+): NavItemState => {
+  const { onClick, value, icon, as, href } = props;
 
   const { selectedValue, onRegister, onUnregister, onSelect } = useNavContext_unstable();
+
+  const rootElementType = as || (href ? 'a' : 'button');
 
   const selected = selectedValue === value;
 
   const innerRef = React.useRef<HTMLElement>(null);
-  const onNavItemClick = useEventCallback(
-    mergeCallbacks(onClick, event => onSelect(event, { type: 'click', event, value })),
+
+  const onNavItemClick: ARIAButtonSlotProps<'a'>['onClick'] = useEventCallback(event => {
+    onClick?.(event);
+
+    if (!event.defaultPrevented && isHTMLElement(event.target)) {
+      onSelect(event, { type: 'click', event, value });
+    }
+  });
+
+  const root = slot.always<ARIAButtonSlotProps<'a'>>(
+    getIntrinsicElementProps(rootElementType, useARIAButtonProps(rootElementType, props)),
+    {
+      elementType: rootElementType,
+      defaultProps: {
+        ref: ref as React.Ref<HTMLButtonElement & HTMLAnchorElement>,
+        type: rootElementType,
+      },
+    },
   );
+
+  root.onClick = onNavItemClick;
 
   React.useEffect(() => {
     onRegister({
@@ -37,17 +62,8 @@ export const useNavItem_unstable = (props: NavItemProps, ref: React.Ref<HTMLAnch
   }, [onRegister, onUnregister, innerRef, value]);
 
   return {
-    components: { root: 'a', icon: 'span' },
-    root: slot.always(
-      getIntrinsicElementProps('a', {
-        ref,
-        role: 'nav',
-        type: 'navigation',
-        ...props,
-        onClick: onNavItemClick,
-      }),
-      { elementType: 'a' },
-    ),
+    components: { root: rootElementType, icon: 'span' },
+    root: root,
     icon: slot.optional(icon, {
       elementType: 'span',
     }),