+++ packages/react-examples/src/react-charting/HorizontalBarChart/HorizontalBarChart.Variant.Example.tsx
@@ -6,6 +6,7 @@ import {
   DataVizPalette,
   getColorFromToken,
 } from '@fluentui/react-charting';
+import { ThemeContext } from '@fluentui/react';
 import { Checkbox } from '@fluentui/react/lib/Checkbox';
 
 interface IHBCVariantExampleState {
@@ -22,6 +23,11 @@ export class HorizontalBarChartVariantExample extends React.Component<{}, IHBCVa
   }
 
   public render() {
+    const IsDarkMode = () => {
+      const theme = React.useContext(ThemeContext);
+      return theme?.isInverted;
+    };
+
     const data: IChartProps[] = [
       {
         chartTitle: 'one',
@@ -29,7 +35,7 @@ export class HorizontalBarChartVariantExample extends React.Component<{}, IHBCVa
           {
             legend: 'one',
             horizontalBarChartdata: { x: 1543, y: 15000 },
-            color: getColorFromToken(DataVizPalette.color17),
+            color: getColorFromToken(DataVizPalette.color17, IsDarkMode()),
           },
         ],
       },
@@ -39,7 +45,7 @@ export class HorizontalBarChartVariantExample extends React.Component<{}, IHBCVa
           {
             legend: 'two',
             horizontalBarChartdata: { x: 800, y: 15000 },
-            color: getColorFromToken(DataVizPalette.color18),
+            color: getColorFromToken(DataVizPalette.color18, IsDarkMode()),
           },
         ],
       },
@@ -49,7 +55,7 @@ export class HorizontalBarChartVariantExample extends React.Component<{}, IHBCVa
           {
             legend: 'three',
             horizontalBarChartdata: { x: 8888, y: 15000 },
-            color: getColorFromToken(DataVizPalette.color19),
+            color: getColorFromToken(DataVizPalette.color19, IsDarkMode()),
           },
         ],
       },
@@ -59,7 +65,7 @@ export class HorizontalBarChartVariantExample extends React.Component<{}, IHBCVa
           {
             legend: 'four',
             horizontalBarChartdata: { x: 15888, y: 15000 },
-            color: getColorFromToken(DataVizPalette.color20),
+            color: getColorFromToken(DataVizPalette.color20, IsDarkMode()),
           },
         ],
       },
@@ -69,7 +75,7 @@ export class HorizontalBarChartVariantExample extends React.Component<{}, IHBCVa
           {
             legend: 'five',
             horizontalBarChartdata: { x: 11444, y: 15000 },
-            color: getColorFromToken(DataVizPalette.color21),
+            color: getColorFromToken(DataVizPalette.color21, IsDarkMode()),
           },
         ],
       },
@@ -79,7 +85,7 @@ export class HorizontalBarChartVariantExample extends React.Component<{}, IHBCVa
           {
             legend: 'six',
             horizontalBarChartdata: { x: 14000, y: 15000 },
-            color: getColorFromToken(DataVizPalette.color22),
+            color: getColorFromToken(DataVizPalette.color22, IsDarkMode()),
           },
         ],
       },
@@ -89,7 +95,7 @@ export class HorizontalBarChartVariantExample extends React.Component<{}, IHBCVa
           {
             legend: 'seven',
             horizontalBarChartdata: { x: 9855, y: 15000 },
-            color: getColorFromToken(DataVizPalette.color23),
+            color: getColorFromToken(DataVizPalette.color23, IsDarkMode()),
           },
         ],
       },
@@ -99,7 +105,7 @@ export class HorizontalBarChartVariantExample extends React.Component<{}, IHBCVa
           {
             legend: 'eight',
             horizontalBarChartdata: { x: 4250, y: 15000 },
-            color: getColorFromToken(DataVizPalette.color24),
+            color: getColorFromToken(DataVizPalette.color24, IsDarkMode()),
           },
         ],
       },