+++ scripts/beachball/release-web-components.config.js
@@ -1,0 +1,19 @@
+require('./register').register();
+
+const { config: sharedConfig } = require('./shared.config');
+const { getConfig } = require('./utils');
+
+const { scope } = getConfig({ version: 'web-components' });
+
+/**
+ * @type {typeof sharedConfig}
+ */
+const config = {
+  ...sharedConfig,
+  scope: [...sharedConfig.scope, ...scope],
+  changelog: {
+    ...sharedConfig.changelog,
+  },
+};
+
+module.exports = config;